done 90%

+++ app/Http/Controllers/JenisBarangController.php
@@ -4,12 +4,16 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\JenisBarang;
-
+use Illuminate\Support\Facades\Auth;
 
 class JenisBarangController extends Controller
 {
     public function index() 
-    { 
+    {   
+        if (Auth::user()->role != 'admin') {
+            return redirect('/')->with('error', 'Unauthorized access');
+        }   
+
         $title = 'Data Jenis-Barang';
         $table_title = 'Table Jenis-Barang';
         $data_jenis_barang = JenisBarang::all();
@@ -17,7 +21,10 @@ class JenisBarangController extends Controller
     }   
 
     public function store(Request $request)
-    {
+    {   
+        if (Auth::user()->role != 'admin') {
+            return redirect('/')->with('error', 'Unauthorized access');
+            }   
         JenisBarang::create([
             'nama_jenis_barang' => $request->nama_jenis_barang,
         ]);
@@ -25,7 +32,10 @@ class JenisBarangController extends Controller
     }   
 
     public function update(Request $request, $id)
-    {
+    {   
+        if (Auth::user()->role != 'admin') {
+            return redirect('/')->with('error', 'Unauthorized access');
+            }   
         JenisBarang::where('id', $id)->where('id', $id)->update([
             'nama_jenis_barang' => $request->nama_jenis_barang,
         ]);
@@ -34,6 +44,9 @@ class JenisBarangController extends Controller
 
     public function destroy($id)
     {   
+        if (Auth::user()->role != 'admin') {
+            return redirect('/')->with('error', 'Unauthorized access');
+            }   
         JenisBarang::where('id', $id)->delete();
         return redirect('/jenisbarang')->with('success', 'Data Berhasil');
     }
